backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
python-multipart==0.0.6
pydantic==2.4.2
openpyxl==3.1.2
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
tensorflow==2.14.0
torch==2.1.0
google-generativeai==0.3.1
python-dotenv==1.0.0
httpx==0.25.0
joblib==1.3.2
